+++ src/types/misc.ts
@@ -1,0 +1,2 @@
+export type Factory<Args, Value> = (args: Args) => Value;
+export type MaybeFactory<Args, Value> = Value | Factory<Args, Value>;